+++ Validator.h
@@ -7,8 +7,8 @@
 //Input Validator
 class Validator {
 public:
-	std::vector<int> GetMoveStringArray(std::string input);
-	bool ValidMove(std::string input, char& pieceType, int& target, int& file, int& rank, char& promotionType , bool& isWhite);
+	std::vector<short> GetMoveStringArray(std::string input);
+	bool ValidMove(std::string input, char& pieceType, short& target, short& file, short& rank, char& promotionType , bool& isWhite);
 private:
 	char pieceChar[5] = { 'N','B','R','Q','K' };
 	char fileChar[8] = { 'a','b','c','d','e','f','g','h' };
@@ -20,11 +20,11 @@ private:
 	bool isTakeChar(char c);
 	bool isCastleChar(char c);
 
-	bool isCastlingMove(std::string input, int& target, bool& isWhite);
-	bool isValidPawnMove(std::string input, int& target, char& promotionType);
-	bool isValidPieceMove(std::string input, int& target, int& file, int& rank);
+	bool isCastlingMove(std::string input, short& target, bool& isWhite);
+	bool isValidPawnMove(std::string input, short& target, char& promotionType);
+	bool isValidPieceMove(std::string input, short& target, short& file, short& rank);
 
-	bool NextTwoLetterIsSquare(std::string input, int nextLetterIndex);
-	bool NextLetterIsEnding(std::string input, int nextLetterIndex);
-	bool NextLetterIsPromotion(std::string input, int nextLetterIndex);
+	bool NextTwoLetterIsSquare(std::string input, short nextLetterIndex);
+	bool NextLetterIsEnding(std::string input, short nextLetterIndex);
+	bool NextLetterIsPromotion(std::string input, short nextLetterIndex);
 };
